cloud_functions/requirements.txt
Flask
gunicorn
google-cloud-storag
google-cloud-vision
google-cloud-aiplatform
functions-framework==3.1.0